requirements.txt
streamlit

python-dotenv==1.0.1
youtube_transcript_api==1.0.2
google-generativeai==0.8.4
